requirements.txt
flask==2.0.1
flask-cas==1.0.2
python-dotenv==0.19.0
gunicorn==20.1.0 